feat: breakout PantryItemCard, update types, general notes
